Extract common behaviors for MemberSearch strategies
